requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
yfinance>=0.2.40
pandas>=2.0.0
pandas-ta>=0.3.14b
requests>=2.31.0
python-dotenv>=1.0.0
